# go-core/docs/swagger.yaml
basePath: /
definitions:
...
  data.MistakeCategory:
    enum:
    - psychological
    - behavioral
    type: string
    x-enum-varnames:
    - MistakeCategoryPsychological
    - MistakeCategoryBehavioral
...
  dto.CreateMistakeRequest:
    properties:
      category:
        $ref: '#/definitions/data.MistakeCategory'
      name:
        maxLength: 255
        minLength: 1
        type: string
      user_id:
        type: integer
    required:
    - category
    - name
    - user_id
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  dto.GetMistakesResponse:
    properties:
      mistakes:
        items:
          $ref: '#/definitions/dto.MistakeResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
...
  dto.MistakeResponse:
    properties:
      category:
        $ref: '#/definitions/data.MistakeCategory'
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.PaginationResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  dto.UpdateMistakeRequest:
    properties:
      category:
        $ref: '#/definitions/data.MistakeCategory'
      id:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      user_id:
        type: integer
    required:
    - category
    - id
    - name
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive trading journal API for tracking trades, setups, and
    market analysis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 100xTrader API
  version: "1.0"
paths:
  /api/v1/mistakes:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of mistakes
      parameters:
      - description: 'Number of mistakes to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of mistakes to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mistakes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetMistakesResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List mistakes
      tags:
      - mistakes
    post:
      consumes:
      - application/json
      description: Create a new trading mistake
      parameters:
      - description: Mistake data
        in: body
        name: mistake
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMistakeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Mistake created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MistakeResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new mistake
      tags:
      - mistakes
  /api/v1/mistakes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific mistake by ID
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mistake deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Invalid mistake ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Mistake not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a mistake
      tags:
      - mistakes
    get:
      consumes:
      - application/json
      description: Retrieve a specific mistake with all its details
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mistake retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MistakeResponse'
              type: object
        "400":
          description: Invalid mistake ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Mistake not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a mistake by ID
      tags:
      - mistakes
    put:
      consumes:
      - application/json
      description: Update an existing mistake with new data
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated mistake data
        in: body
        name: mistake
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMistakeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Mistake updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MistakeResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Mistake not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a mistake
      tags:
      - mistakes
  /api/v1/mistakes/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of mistakes for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 'Number of mistakes to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of mistakes to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User mistakes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetMistakesResponse'
              type: object
        "400":
          description: Invalid user ID or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get mistakes by user
      tags:
      - mistakes
  /api/v1/mistakes/user/{user_id}/category/{category}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of mistakes filtered by category for
        a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Mistake category
        in: path
        name: category
        required: true
        type: string
      - description: 'Number of mistakes to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of mistakes to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User mistakes by category retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetMistakesResponse'
              type: object
        "400":
          description: Invalid user ID, category, or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get mistakes by category
      tags:
      - mistakes
...
schemes:
- http
swagger: "2.0"